Week-3/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.2
    # via streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   typer-slim
    #   uvicorn
    #   wandb
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
coverage[toml]==7.12.0
    # via pytest-cov
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.3.1
    # via
    #   anyio
    #   pytest
fastapi==0.123.2
    # via -r requirements.in
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
fonttools==4.61.0
    # via matplotlib
fsspec==2025.10.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via
    #   streamlit
    #   wandb
grad-cam==1.5.5
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   huggingface-hub
huggingface-hub==1.1.7
    # via timm
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.2
    # via scikit-learn
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
lightning-utilities==0.15.2
    # via torchmetrics
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.7
    # via
    #   -r requirements.in
    #   grad-cam
    #   seaborn
mpmath==1.3.0
    # via sympy
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.13.0
    # via altair
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   grad-cam
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   seaborn
    #   streamlit
    #   torchmetrics
    #   torchvision
opencv-python==4.12.0.88
    # via grad-cam
packaging==25.0
    # via
    #   altair
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   pandera
    #   pytest
    #   streamlit
    #   torchmetrics
    #   wandb
pandas==2.3.3
    # via
    #   -r requirements.in
    #   seaborn
    #   streamlit
pandera==0.27.0
    # via -r requirements.in
pillow==12.0.0
    # via
    #   grad-cam
    #   matplotlib
    #   streamlit
    #   torchvision
platformdirs==4.5.0
    # via wandb
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
protobuf==6.33.1
    # via
    #   streamlit
    #   wandb
pyarrow==21.0.0
    # via streamlit
pydantic==2.12.5
    # via
    #   fastapi
    #   pandera
    #   wandb
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pyparsing==3.2.5
    # via matplotlib
pyside6==6.10.1
    # via -r requirements.in
pyside6-addons==6.10.1
    # via pyside6
pyside6-essentials==6.10.1
    # via
    #   pyside6
    #   pyside6-addons
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==7.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   timm
    #   uvicorn
    #   wandb
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   streamlit
    #   wandb
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.7.0
    # via timm
scikit-learn==1.7.2
    # via
    #   -r requirements.in
    #   grad-cam
scipy==1.15.3
    # via scikit-learn
seaborn==0.13.2
    # via -r requirements.in
sentry-sdk==2.46.0
    # via wandb
shellingham==1.5.4
    # via huggingface-hub
shiboken6==6.10.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
starlette==0.50.0
    # via fastapi
streamlit==1.51.0
    # via -r requirements.in
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
timm==1.0.22
    # via -r requirements.in
toml==0.10.2
    # via streamlit
tomli==2.3.0
    # via
    #   coverage
    #   pytest
torch==2.9.1
    # via
    #   -r requirements.in
    #   grad-cam
    #   timm
    #   torchmetrics
    #   torchvision
torchinfo==1.8.0
    # via -r requirements.in
torchmetrics==1.8.2
    # via -r requirements.in
torchvision==0.24.1
    # via
    #   -r requirements.in
    #   grad-cam
    #   timm
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   grad-cam
    #   huggingface-hub
ttach==0.0.3
    # via grad-cam
typeguard==4.4.4
    # via pandera
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   huggingface-hub
    #   lightning-utilities
    #   pandera
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   streamlit
    #   torch
    #   typeguard
    #   typer-slim
    #   typing-inspect
    #   typing-inspection
    #   uvicorn
    #   wandb
typing-inspect==0.9.0
    # via pandera
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.38.0
    # via -r requirements.in
wandb==0.23.0
    # via -r requirements.in
watchdog==6.0.0
    # via streamlit
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
